% Generated by roxygen2 (4.0.2): do not edit by hand
\name{attribute.instances.metadata.info}
\alias{attribute.instances.metadata.info}
\title{List of tables containing one of given attributes}
\usage{
\method{attribute.instances}{metadata.info}(metadata.info, attribute.names,
  fixed = TRUE, partial = FALSE)
}
\arguments{
\item{metadata.info}{an object of \code{\link{metadata.info}} class}

\item{attribute.names}{character vector of attribute names}

\item{fixed}{should the attribute names be treated as fixed strings (default) or regular expressions?}

\item{partial}{should the attribute names be matched partially (e.g. attribute \code{"attrib123"} will match query \code{"attrib12"}) or fully?}
}
\description{
Extracts from an instance of \code{\link{metadata.info}} object
all occurences of attributes from \code{attribute.names}
}

