% Generated by roxygen2 (4.0.2): do not edit by hand
\name{db.attribute.pg.constraints}
\alias{db.attribute.pg.constraints}
\title{Constructor for class db.attribute.pg.constraints}
\usage{
db.attribute.pg.constraints(schemaname, tablename, attname, notnull = FALSE,
  hasdefault = FALSE, internalid, internalname)
}
\arguments{
\item{schemaname}{the name of DB schema}

\item{tablename}{the name of DB table}

\item{attname}{the name of the attribute}

\item{notnull}{has the attribute NOT NULL constraint?}

\item{hasdefault}{has the attribute default value?}

\item{internalid}{(reserved)}

\item{internalname}{(reserved)}
}
\description{
PostgreSQL constraints concerning the attribute
}

