% Generated by roxygen2 (4.0.2): do not edit by hand
\name{db.attributes.counts.metadata.info}
\alias{db.attributes.counts.metadata.info}
\title{DB attributes}
\usage{
\method{db.attributes}{counts.metadata.info}(metadata.inf, typeinfo = FALSE)
}
\arguments{
\item{metadata.inf}{object of class \code{\link{metadata.info}} to perform extraction from}

\item{typeinfo}{should the DB type be included?}
}
\description{
Extracts the DB attributes from \code{\link{metadata.info}} counting the instances
}

