% Generated by roxygen2 (4.0.2): do not edit by hand
\name{db.attributes.metadata.info}
\alias{db.attributes.metadata.info}
\title{DB attributes}
\usage{
\method{db.attributes}{metadata.info}(metadata.inf, full = FALSE)
}
\arguments{
\item{metadata.inf}{object of class \code{\link{metadata.info}} to perform extraction from}

\item{full}{should full database information be returned?}
}
\description{
Extracts the DB attributes from \code{\link{metadata.info}} object
}
\details{
Depending on the value of \code{\link{full}} parameter the function returns
\code{data.frame} describing each attribute in \code{\link{metadata.info}} object.

Default (\code{\link{full} = FALSE}) columns of returned object:
\itemize{
   \item{schemaname}
   \item{tablename}
   \item{attribute name}
   \item{type of attribute}
}

More detailed information (\code{\link{full} = TRUE}) includes default columns, and additionally:
\itemize{
   \item{length of attribute (related to type)}
   \item{estimated fraction of NULL values}
   \item{estimated number of distinct values (caution!)}
   \item{most common values}
   \item{frequency of most common values}
   \item{histogram bounds}
}
}

