% Generated by roxygen2 (4.0.2): do not edit by hand
\name{db.connection}
\alias{db.connection}
\title{DB connection specification}
\usage{
db.connection(...)
}
\description{
DB connection specification
}
\examples{
control.connection <- db.connection(
 user = "dbuser",
 host = "dbhost",
 port = 0000,
 dbname = "dbname",
 password = "dbpass"
 )
}

