% Generated by roxygen2 (4.0.2): do not edit by hand
\name{db.tables.metadata.info}
\alias{db.tables.metadata.info}
\title{DB tables}
\usage{
\method{db.tables}{metadata.info}(metadata.inf)
}
\arguments{
\item{metadata.inf}{object of class \code{\link{metadata.info}} to perform extraction from}
}
\description{
Extracts the DB tables from \code{\link{metadata.info}} object
}

