% Generated by roxygen2 (4.0.2): do not edit by hand
\name{describe.attribute.metadata.info}
\alias{describe.attribute.metadata.info}
\title{Attribute description}
\usage{
\method{describe.attribute}{metadata.info}(metadata.info, attribute.name)
}
\arguments{
\item{metadata.info}{an object of \code{\link{metadata.info}} class}

\item{attribute.name}{the attribute name to describe}
}
\description{
Attribute description
}

