% Generated by roxygen2 (4.0.2): do not edit by hand
\name{describe.table.metadata.info}
\alias{describe.table.metadata.info}
\title{Table description}
\usage{
\method{describe.table}{metadata.info}(metadata.inf, table.name)
}
\arguments{
\item{metadata.inf}{an object of \code{\link{metadata.info}} class}

\item{table.name}{the table name to describe}
}
\description{
Table description
}

