% Generated by roxygen2 (4.0.2): do not edit by hand
\name{experimental.db.sketch.attribute}
\alias{experimental.db.sketch.attribute}
\title{Multiselect for given attribute}
\usage{
experimental.db.sketch.attribute(db.connection, metadata.inf, attribute.name,
  attribute.value, results.limit = 10)
}
\arguments{
\item{db.connection}{object of S3 class \code{\link{db.connection}}}

\item{metadata.inf}{object of S3 class \code{\link{metadata.info}}}

\item{attribute.name}{(\code{character}) the name of the attribute}

\item{attribute.value}{(\code{character}) a value of the attribute}

\item{results.limit}{(\code{integer}) a maximal number of returned rows}
}
\description{
Function performs a multiple selects from tables (in \code{metadata.inf}) containing the attribute \code{attribute.name}:
\itemize{
 \item counts all records of attribute \code{attribute.name} with value \code{attribute.value}
 \item selects a sample of at most \code{results.limit} rows
}
}

