% Generated by roxygen2 (4.0.2): do not edit by hand
\name{experimental.infer.domains}
\alias{experimental.infer.domains}
\title{Inference of attribute domains}
\usage{
experimental.infer.domains(meta.inf, minimal.support = 0.1,
  maximal.length = 20)
}
\arguments{
\item{minimal.support}{the minimal support (among the tables) for the domain}

\item{maximal.length}{the maximal number of attributes in the domain}

\item{metadata.inf}{object of S3 class \code{\link{metadata.info}}}
}
\value{
object of class \code{\link[arules]{itemsets}} from the \emph{arules} package
}
\description{
Function searches for the frequently co-occuring sets of attributes
}

