% Generated by roxygen2 (4.0.2): do not edit by hand
\name{experimental.infer.domains2}
\alias{experimental.infer.domains2}
\title{Inference of attribute domains}
\usage{
experimental.infer.domains2(meta.inf, domain.set = list(),
  minimal.support = 0.1, minimal.length = 2, maximal.length = 20)
}
\arguments{
\item{domain.set}{set of domains}

\item{minimal.support}{the minimal support (among the tables) for the domain}

\item{maximal.length}{the maximal number of attributes in the domain}

\item{metadata.inf}{object of S3 class \code{\link{metadata.info}}}
}
\value{
object of class \code{\link[arules]{itemsets}} from the \emph{arules} package
}
\description{
Function searches for the frequently co-occuring sets of attributes
}

