% Generated by roxygen2 (4.0.2): do not edit by hand
\name{get.tables.query}
\alias{get.tables.query}
\title{Table list and simple metadata extraction}
\usage{
get.tables.query(schemaname = NULL)
}
\arguments{
\item{schemaname}{name of specific schema, if not provided tables from all schemas will be returned}
}
\value{
SQL query retrieving the metadata information concerning the tables
}
\description{
Constructs the query retrieving the metadata information concerning the tables:
\itemize{
   \item{schemaname}
   \item{tablename}
   \item{number of attributes}
   \item{estimate of counts}
   \item{is table indexed?}
   \item{has table a primary key?}
}
}

