% Generated by roxygen2 (4.0.2): do not edit by hand
\name{load.metadata.anonymized}
\alias{load.metadata.anonymized}
\title{Metadata anonymizing query}
\usage{
load.metadata.anonymized(control.connection, schemaname = NULL)
}
\arguments{
\item{control.connection}{the \code{\link{db.connection}} object}
}
\value{
instance of \code{\link{metadata.info}} class
}
\description{
Loads the attributes list, table list and make anonymization
}

