% Generated by roxygen2 (4.0.2): do not edit by hand
\name{metadata.info}
\alias{metadata.info}
\title{Metadata S3 class}
\usage{
metadata.info(tables, attributes, schema = NULL)
}
\arguments{
\item{tables}{result of \code{\link{sql.tables}} query}

\item{attributes}{result of \code{\link{sql.attributes}} query}

\item{schema}{the schema(s) restricting the queries}
}
\description{
Metadata S3 class
}

