% Generated by roxygen2 (4.0.2): do not edit by hand
\name{query.flatten.results}
\alias{query.flatten.results}
\title{Flattens results of the query.load.execute to a single data frame}
\usage{
query.flatten.results(query.load.execute.results,
  idfield.proposed = "query.flatten.results.id")
}
\arguments{
\item{query.load.execute.results}{the results to be processed}

\item{idfield.proposed}{the proposed name of the column storing the row name (extracted from \code{query.load.execute.results})}
}
\value{
\code{list} with following fields:
 \itemize{
   \item{\code{results}} the resulting data.frame
   \item{\code{idfield}} the name of the column in \code{results} preserving the list name of the query (may differ from \code{idfield.proposed} in case of name collision)
 }
}
\description{
Flattens results of the query.load.execute to a single data frame, if results do match the same structure (column names)
}

