% Generated by roxygen2 (4.0.2): do not edit by hand
\name{query.load.execute}
\alias{query.load.execute}
\title{Query execution}
\usage{
query.load.execute(queries, control.connection, verbose = TRUE)
}
\arguments{
\item{queries}{list of queries to execute}

\item{control.connection}{the object of \code{db.connection} class}

\item{verbose}{should additional information be printed out}
}
\description{
Executes the list of queries providing a list of dataframes equipped with the attributes:
\itemize{
 \item{data.sqlQuery (the query used for constructing the dataframe)}
 \item{data.retirevedOn (the time the query returned values)}
}
}

