% Generated by roxygen2 (4.0.2): do not edit by hand
\name{restrict.metadata.info}
\alias{restrict.metadata.info}
\title{Filtering the metadata.info}
\usage{
\method{restrict}{metadata.info}(meta.inf, restr.obj)
}
\arguments{
\item{meta.inf}{the \code{\link{metadata.info}} object to be restricted}

\item{restr.obj}{description of the restriction}
}
\value{
restricted \code{\link{metadata.info}} object
}
\description{
The function provides a mechanism for restricting the \code{\link{metadata.info}} object.
}
\details{
It can be provided with:
\itemize{
 \item{subset of rows or columns from the result of \code{\link{db.attributes}}} function
 \item{subset of rows or columns from the result of \code{\link{db.tables}}} function
}
As a result new  \code{\link{metadata.info}} object is created matching the filtering criteria
}
\examples{
db.atts <- db.attributes(meta.inf)
# let's select attributes with more than 100 instances
db.atts.cnt <- db.attributes.counts(meta.inf)  \%>\% dplyr::filter(tablecount>100)
meta.inf.restr <- restrict(meta.inf, db.atts.cnt)
}

