% Generated by roxygen2 (4.0.2): do not edit by hand
\name{sql.attribute.characteristics}
\alias{sql.attribute.characteristics}
\title{Attribute characteristics}
\usage{
sql.attribute.characteristics(query, query.colname, query.countname,
  query.nullitymarker)
}
\arguments{
\item{query}{the query to analyze}

\item{query.colname}{the attribute name}

\item{query.countname}{the alias for counting variable (some valid SQL name not among the attributes generated by query)}

\item{query.nullitymarker}{the alias for counting variable (some valid SQL name not among the attributes generated by query)}
}
\description{
Returns the query returning:
\itemize{
 \item row_cnt number of rows returned by \code{query}
 \item distinct_val_cnt number of distinct values of \code{query.colname}
 \item mincnt minimal number of instances for some distinct value of \code{query.colname}
 \item avgcnt average number of instances of distinct values of \code{query.colname}
 \item maxcnt maximal number of instances for some distinct value of \code{query.colname}
 \item sdevcnt standard deviation of the number of distinct value instances \code{query.colname}
 \item null_cnt number of NULL instances of \code{query.colname}
}
}

