% Generated by roxygen2 (4.0.2): do not edit by hand
\name{sql.attributes}
\alias{sql.attributes}
\title{Table list and simple metadata extraction}
\usage{
sql.attributes(schemaname = NULL)
}
\arguments{
\item{schemaname}{name of specific schema, if not provided tables from all schemas will be returned}
}
\description{
Constructs the query retrieving the metadata information of each attribute in each table as well as statistics computed by PostreSQL:
\itemize{
   \item{schemaname}
   \item{tablename}
   \item{attribute name}
   \item{type of attribute}
   \item{length of attribute (related to type)}
   \item{estimated fraction of NULL values}
   \item{estimated number of distinct values (caution!)}
   \item{most common values}
   \item{frequency of most common values}
   \item{histogram bounds}
}
}

