% Generated by roxygen2 (4.0.2): do not edit by hand
\name{sql.fill.template}
\alias{sql.fill.template}
\title{Generic template filler}
\usage{
sql.fill.template(template.name, query.names, param.list)
}
\arguments{
\item{template.name}{the name of the template}

\item{query.names}{the vector of names of the generated queries (in the genereated query list)}

\item{param.list}{the (named) list with names of parameters to be replaced and the values to be substituted}
}
\description{
Generic template filler
}
\examples{
sql.fill.template("dictionary_values", query.names="q1", list("\%\%DICTKEY\%\%"="abd", "\%\%QUERY\%\%"="bca"))
 sql.fill.template("dictionary_values", query.names=c("q1","q2"), list("\%\%QUERY\%\%"=c("bca","bcb"), "\%\%DICTKEY\%\%"=c("abd")))
 sql.fill.template("dictionary_values", query.names=c("q1","q2"), list("\%\%QUERY\%\%"=c("bca", "bcb"), "\%\%DICTKEY\%\%"=c("abc","abd")))
}

