% Generated by roxygen2 (4.0.2): do not edit by hand
\name{sql.table.schemed}
\alias{sql.table.schemed}
\title{SQL formation: combining schema name with table}
\usage{
sql.table.schemed(tab.name, schema.name)
}
\arguments{
\item{tab.name}{(\code{character} or \code{character} vector) name(s) of the tables}

\item{schema.name}{(\code{character} or \code{character} vector) name(s) of the schema(s)}
}
\description{
For each element of \code{tab.name} the corresponding element from \code{schema.name}
is used for forming the valid table reference in format \code{schema.table}.
If the corresponding element of the \code{schema.name} is \code{NA} or empty (consisting of white chars only)
the table reference is simply \code{table}.
}

