% Generated by roxygen2 (4.0.2): do not edit by hand
\name{sql.template.fill}
\alias{sql.template.fill}
\title{Generic SQL template filler}
\usage{
sql.template.fill(template.repo, template.name, query.names, param.list)
}
\arguments{
\item{template.repo}{repository of SQL templates}

\item{template.name}{the name of the template}

\item{query.names}{the vector of names of the generated queries (in the genereated query list)}

\item{param.list}{the (named) list with names of parameters to be replaced and the values to be substituted}
}
\description{
Generic SQL template filler
}
\examples{
sql.template.fill(
   some.repo,
   "dictionary_values",
   query.names="q1",
   list("\%\%DICTKEY\%\%" = "abd", "\%\%QUERY\%\%" = "bca")
   )

 sql.template.fill(
   some.repo,
   "dictionary_values",
   query.names = c("q1","q2"),
   list("\%\%QUERY\%\%" = c("bca", "bcb"), "\%\%DICTKEY\%\%" = c("abd"))
   )

 sql.template.fill(
   some.repo,
   "dictionary_values",
   query.names = c("q1", "q2"),
   list("\%\%QUERY\%\%" = c("bca", "bcb"), "\%\%DICTKEY\%\%"=c("abc", "abd"))
   )
}

