% Generated by roxygen2 (4.0.2): do not edit by hand
\name{string.get.most.similar}
\alias{string.get.most.similar}
\title{'Fuzzy' string matching}
\usage{
string.get.most.similar(source.string, similarityset)
}
\arguments{
\item{source.string}{the string for which similarity is to be calculated}

\item{similarityset}{vector of strings to be matched}
}
\description{
'Fuzzy' string matching
}

