% Generated by roxygen2 (4.0.2): do not edit by hand
\name{tables.with.attributes.metadata.info}
\alias{tables.with.attributes.metadata.info}
\title{List of tables with containing all attributes from given set}
\usage{
\method{tables.with.attributes}{metadata.info}(metadata.info, attribute.names)
}
\arguments{
\item{metadata.info}{an object of metadata.info class}

\item{attribute.names}{character vector of attribute names}
}
\description{
Exctracts from an instance of \code{metadata.info} a vector of tables
containing all attributes from \code{attribute.names}
}

