% Generated by roxygen2 (4.0.2): do not edit by hand
\name{attribute.instances.metadata.info}
\alias{attribute.instances.metadata.info}
\title{List of tables containing one of given attributes}
\usage{
\method{attribute.instances}{metadata.info}(metadata.info, attribute.names)
}
\arguments{
\item{metadata.info}{an object of metadata.info class}

\item{attribute.names}{character vector of attribute names}
}
\description{
Extracts from an instance of \code{metadata.info} object
all occurences of attributes from \code{attribute.names}
}

