% Generated by roxygen2 (4.0.2): do not edit by hand
\name{db.attribute}
\alias{db.attribute}
\title{Constructor for class db.attribute}
\usage{
db.attribute(schemaname, tablename, attname, typename, attlen = -1,
  attlen2 = -1, colposition, internalid, internalname)
}
\arguments{
\item{schemaname}{the name of DB schema}

\item{tablename}{the name of DB table}

\item{attname}{the name of the attribute}

\item{attlen}{primary length of the type}

\item{attlen2}{secondary length of the type}

\item{colposition}{position of the column in table}

\item{internalid}{(reserved)}

\item{internalname}{(reserved)}
}
\description{
Description of PostgreSQL attribute
}

