% Generated by roxygen2 (4.0.2): do not edit by hand
\name{load.metadata}
\alias{load.metadata}
\title{Metadata query}
\usage{
load.metadata(control.connection, schemaname = NULL)
}
\arguments{
\item{control.connection}{the control.connection object}
}
\description{
Loads the attributes list and table list
}

